functions/requirements.txt
firebase_functions~=0.1.0
firebase_admin~=6.4.0
google-genai~=1.0.0
Pillow~=10.4.0
